@@ -10,7 +10,6 @@ using ApplicationCore.Middlewares;
 using AutoMapper;
 using Hangfire;
 using Hangfire.SqlServer;
-using IdentityServer4.Services;
 using Infrastructure.Caches;
 using Infrastructure.Data;
 using Infrastructure.Repositories;
@@ -37,21 +36,19 @@ using TMTDentalAPI.Middlewares;
 using TMTDentalAPI.Services;
 using Umbraco.Web.Mapping;
 using TMTDentalAPI.JobFilters;
+using Microsoft.Extensions.Hosting;
+using Microsoft.OpenApi.Models;
 
 namespace TMTDentalAPI
 {
     public class Startup
     {
-        private readonly ILoggerFactory _loggerFactory;
-        public Startup(IConfiguration configuration, IHostingEnvironment environment, ILoggerFactory loggerFactory)
+        public Startup(IConfiguration configuration)
         {
             Configuration = configuration;
-            Environment = environment;
-            _loggerFactory = loggerFactory;
         }
 
         public IConfiguration Configuration { get; }
-        public IHostingEnvironment Environment { get; }
 
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
@@ -293,8 +290,7 @@ namespace TMTDentalAPI
                 {
                     builder.AllowAnyOrigin()
                     .AllowAnyMethod()
-                    .AllowAnyHeader()
-                    .AllowCredentials();
+                    .AllowAnyHeader();
                 });
             });
             services.AddCors();
@@ -324,26 +320,39 @@ namespace TMTDentalAPI
             //GlobalJobFilters.Filters.Add(new ServerTenantFilter());
             //GlobalJobFilters.Filters.Add(new ClientTenantFilter());
 
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            services.AddControllersWithViews();
 
             services.AddSwaggerGen(c =>
             {
-                c.SwaggerDoc("v1", new Info { Title = "My API", Version = "v1" });
+                c.SwaggerDoc("v1", new OpenApiInfo { Title = "My API", Version = "v1" });
 
-                // Swagger 2.+ support
-                var security = new Dictionary<string, IEnumerable<string>>
-                {
-                    {"Bearer", new string[] { }},
-                };
-
-                c.AddSecurityDefinition("Bearer", new ApiKeyScheme
+                c.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
                 {
                     Description = "Standard Authorization header using the Bearer scheme. Example: \"bearer {token}\"",
-                    In = "header",
+                    In = ParameterLocation.Header,
                     Name = "Authorization",
-                    Type = "apiKey"
+                    Type = SecuritySchemeType.ApiKey,
+                    Scheme = "Bearer"
                 });
-                c.AddSecurityRequirement(security);
+
+                c.AddSecurityRequirement(new OpenApiSecurityRequirement()
+                {
+                    {
+                        new OpenApiSecurityScheme
+                        {
+                            Reference = new OpenApiReference
+                            {
+                                Type = ReferenceType.SecurityScheme,
+                                Id = "Bearer"
+                            },
+                            Scheme = "oauth2",
+                            Name = "Bearer",
+                            In = ParameterLocation.Header,
+
+                        },
+                        new List<string>()
+                    }
+                });
             });
 
 
@@ -359,7 +368,7 @@ namespace TMTDentalAPI
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -372,11 +381,16 @@ namespace TMTDentalAPI
                 app.UseHsts();
             }
 
-            app.UseMultitenancy<AppTenant>();
-
             app.UseHttpsRedirection();
             app.UseStaticFiles();
-            app.UseSpaStaticFiles();
+            if (!env.IsDevelopment())
+            {
+                app.UseSpaStaticFiles();
+            }
+
+            app.UseRouting();
+
+            app.UseMultitenancy<AppTenant>();
 
             // Enable middleware to serve generated Swagger as a JSON endpoint.
             app.UseSwagger();
@@ -392,22 +406,20 @@ namespace TMTDentalAPI
             app.UseCors(
                 options => options.WithOrigins("http://localhost:4200").AllowAnyMethod().AllowAnyHeader().AllowCredentials()
             );
+
             app.UseMiddleware<GetTokenFromQueryStringMiddleware>();
-            app.UseAuthentication();
             app.UseMiddleware<MigrateDbMiddleware>();
             app.UseMiddleware(typeof(ErrorHandlingMiddleware));
 
-            app.UseSignalR(routes =>
-            {
-                routes.MapHub<NotificationHub>("/notification");
-                routes.MapHub<AppointmentHub>("/appointment");
-            });
+            app.UseAuthentication();
 
-            app.UseMvc(routes =>
+            app.UseAuthorization();
+
+            app.UseEndpoints(endpoints =>
             {
-                routes.MapRoute(
+                endpoints.MapControllerRoute(
                     name: "default",
-                    template: "{controller}/{action=Index}/{id?}");
+                    pattern: "{controller}/{action=Index}/{id?}");
             });
 
             app.UseSpa(spa =>
