@@ -426,7 +426,7 @@ export class SaleOrderCreateUpdateComponent implements OnInit {
 
   createRecord() {
     var val = this.formGroup.value;
-    val.dateOrder = this.intlService.formatDate(val.dateOrderObj, 'g', 'en-US');
+    val.dateOrder = this.intlService.formatDate(val.dateOrderObj, 'yyyy-MM-ddTHH:mm:ss');
     val.partnerId = val.partner.id;
     val.pricelistId = val.pricelist.id;
     val.userId = val.user ? val.user.id : null;
@@ -439,7 +439,7 @@ export class SaleOrderCreateUpdateComponent implements OnInit {
 
   saveRecord() {
     var val = this.formGroup.value;
-    val.dateOrder = this.intlService.formatDate(val.dateOrderObj, 'g', 'en-US');
+    val.dateOrder = this.intlService.formatDate(val.dateOrderObj, 'yyyy-MM-ddTHH:mm:ss');
     val.partnerId = val.partner.id;
     val.pricelistId = val.pricelist.id;
     val.userId = val.user ? val.user.id : null;
@@ -615,7 +615,7 @@ export class SaleOrderCreateUpdateComponent implements OnInit {
 
   getFormDataSave() {
     var val = this.formGroup.value;
-    val.dateOrder = this.intlService.formatDate(val.dateOrderObj, 'g', 'en-US');
+    val.dateOrder = this.intlService.formatDate(val.dateOrderObj, 'yyyy-MM-ddTHH:mm:ss');
     val.partnerId = val.partner.id;
     val.pricelistId = val.pricelist.id;
     val.userId = val.user ? val.user.id : null;
@@ -716,7 +716,7 @@ export class SaleOrderCreateUpdateComponent implements OnInit {
     }
 
     var val = this.formGroup.value;
-    val.dateOrder = this.intlService.formatDate(val.dateOrderObj, 'g', 'en-US');
+    val.dateOrder = this.intlService.formatDate(val.dateOrderObj, 'yyyy-MM-ddTHH:mm:ss');
     val.partnerId = val.partner.id;
     val.pricelistId = val.pricelist.id;
     val.userId = val.user ? val.user.id : null;
